requirements.txt
fastapi
uvicorn
langchain
openai
python-dotenv
pydantic
langchain_community
